requirements.txt
asgiref==3.8.1
bleach==6.2.0
boto3==1.36.9
botocore==1.36.9
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==4.2.18
django-countries==7.6.1
django-crispy-forms==2.3
django-js-asset==3.0.1
django-storages==1.14.4
django-summernote==0.8.20.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
sentry-sdk==2.20.0
six==1.17.0
sqlparse==0.5.3
stripe==11.3.0b2
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
webencodings==0.5.1
